mcp-servers/lvgl-sim/requirements.txt
mcp>=1.0.0
Pillow>=10.0.0
click>=8.0.0
pytest>=7.0.0
